+++ ICanSeeClearlyNow/Config/PluginConfig.cs
@@ -5,6 +5,7 @@ namespace ICanSeeClearlyNow {
   public class PluginConfig {
     public static ConfigEntry<bool> IsModEnabled { get; private set; }
     public static ConfigEntry<float> GammaValue { get; private set; }
+    public static ConfigEntry<bool> IsAshlandsAlwaysOn {  get; private set; }
     public static ConfigEntry<bool> IsCustomTimer { get; private set; }
     public static ConfigEntry<float> IsCustomNight {  get; private set; }
     public static ConfigEntry<float> IsCustomDay { get; private set; }
@@ -14,6 +15,8 @@ namespace ICanSeeClearlyNow {
 
       GammaValue = Config.BindInOrder("Gamma Setting", "GammaValue", defaultValue: 0.6f, "Set your Gamma Value between 0.3 and 0.99",
           new AcceptableValueRange<float>(0.3f, 0.99f));
+
+      IsAshlandsAlwaysOn = Config.BindInOrder<bool>("Always On Options", "IsAshlandsAlwaysOn", false, "Always have Gamma Adjustment in the Ashlands");
 
       IsCustomTimer = Config.BindInOrder<bool>("Custom Time", "IsCustomTimer", false, "Enable or Disable custom Night/Day Values");