+++ ICanSeeClearlyNow/Patches/GammaCorrection.cs
@@ -1,0 +1,78 @@
+﻿using HarmonyLib;
+
+using UnityEngine;
+
+using static ICanSeeClearlyNow.PluginConfig;
+
+namespace ICanSeeClearlyNow
+{
+  [HarmonyPatch]
+
+  public class AmbientLightPatch
+  {
+
+    static Color originalColor;
+
+    [HarmonyPrefix]
+    [HarmonyPatch(typeof(RenderSettings), "set_ambientLight")]
+    static void SetLightPrefix(ref Color value)
+    {
+      if (!IsModEnabled.Value || EnvMan.instance == null)
+      {
+        return;
+      }
+
+      originalColor = value;
+
+      value = LightenColor(value);
+    }
+
+    [HarmonyPostfix]
+    [HarmonyPatch(typeof(RenderSettings), "get_ambientLight")]
+    static void GetLight(ref Color __result)
+    {
+      if (!IsModEnabled.Value)
+      {
+        return;
+      }
+      __result = originalColor;
+    }
+
+    static Color LightenColor(Color sourceColor)
+    {
+
+      float avgColor = (sourceColor.r + sourceColor.g + sourceColor.b) / 3;
+
+      bool isNight = false;
+      bool isAshlands = EnvMan.instance.GetBiome() == Heightmap.Biome.AshLands;
+
+      if (IsAshlandsAlwaysOn.Value && isAshlands)
+      {
+        sourceColor *= GammaValue.Value / avgColor;
+      }
+
+      if (IsCustomTimer.Value)
+      {
+        float timeofday = EnvMan.instance.GetDayFraction();
+        float icn = IsCustomNight.Value / 24f;
+        float icd = IsCustomDay.Value / 24f;
+
+        isNight = timeofday >= icn || timeofday <= icd || EnvMan.IsNight() || EnvMan.instance.GetCurrentEnvironment().m_alwaysDark;
+
+        if (isNight)
+        {
+          sourceColor *= GammaValue.Value / avgColor;
+        }
+      }
+
+      if (!IsCustomTimer.Value)
+      {
+        if (EnvMan.IsNight() || EnvMan.instance.GetCurrentEnvironment().m_alwaysDark)
+        {
+          sourceColor *= GammaValue.Value / avgColor;
+        }
+      }
+      return sourceColor;
+    }
+  }
+}